+++ src/aac_metrics/classes/coco_meteor.py
@@ -26,29 +26,29 @@ class CocoMETEOR(Metric):
         verbose: int = 0,
     ) -> None:
         super().__init__()
-        self.return_all_scores = return_all_scores
-        self.java_path = java_path
-        self.cache_path = cache_path
-        self.java_max_memory = java_max_memory
-        self.verbose = verbose
+        self._return_all_scores = return_all_scores
+        self._java_path = java_path
+        self._cache_path = cache_path
+        self._java_max_memory = java_max_memory
+        self._verbose = verbose
 
-        self.candidates = []
-        self.mult_references = []
+        self._candidates = []
+        self._mult_references = []
 
     def compute(self) -> Union[Tensor, tuple[dict[str, Tensor], dict[str, Tensor]]]:
         return coco_meteor(
-            self.candidates,
-            self.mult_references,
-            self.return_all_scores,
-            self.java_path,
-            self.cache_path,
-            self.java_max_memory,
-            self.verbose,
+            self._candidates,
+            self._mult_references,
+            self._return_all_scores,
+            self._java_path,
+            self._cache_path,
+            self._java_max_memory,
+            self._verbose,
         )
 
     def reset(self) -> None:
-        self.candidates = []
-        self.mult_references = []
+        self._candidates = []
+        self._mult_references = []
         return super().reset()
 
     def update(
@@ -56,5 +56,5 @@ class CocoMETEOR(Metric):
         candidates: list[str],
         mult_references: list[list[str]],
     ) -> None:
-        self.candidates += candidates
-        self.mult_references += mult_references
+        self._candidates += candidates
+        self._mult_references += mult_references