+++ src/aac_metrics/classes/evaluate.py
@@ -8,7 +8,7 @@ from typing import Callable, Iterable, Union
 from torch import Tensor
 
 from aac_metrics.functional.evaluate import evaluate, _get_metrics_set
-from aac_metrics.modules.base import Metric
+from aac_metrics.classes.base import Metric
 
 
 logger = logging.getLogger(__name__)
@@ -39,30 +39,30 @@ class Evaluate(Metric, list[Metric]):
 
         Metric.__init__(self)
         list.__init__(self, metrics)
-        self.use_ptb_tokenizer = use_ptb_tokenizer
-        self.java_path = java_path
-        self.tmp_path = tmp_path
-        self.cache_path = cache_path
-        self.verbose = verbose
+        self._use_ptb_tokenizer = use_ptb_tokenizer
+        self._java_path = java_path
+        self._tmp_path = tmp_path
+        self._cache_path = cache_path
+        self._verbose = verbose
 
-        self.candidates = []
-        self.mult_references = []
+        self._candidates = []
+        self._mult_references = []
 
     def compute(self) -> tuple[dict[str, Tensor], dict[str, Tensor]]:
         return evaluate(
-            self.candidates,
-            self.mult_references,
-            self.use_ptb_tokenizer,
+            self._candidates,
+            self._mult_references,
+            self._use_ptb_tokenizer,
             self,
-            java_path=self.java_path,
-            tmp_path=self.tmp_path,
-            cache_path=self.cache_path,
-            verbose=self.verbose,
+            java_path=self._java_path,
+            tmp_path=self._tmp_path,
+            cache_path=self._cache_path,
+            verbose=self._verbose,
         )
 
     def reset(self) -> None:
-        self.candidates = []
-        self.mult_references = []
+        self._candidates = []
+        self._mult_references = []
         return super().reset()
 
     def update(
@@ -70,5 +70,5 @@ class Evaluate(Metric, list[Metric]):
         candidates: list[str],
         mult_references: list[list[str]],
     ) -> None:
-        self.candidates += candidates
-        self.mult_references += mult_references
+        self._candidates += candidates
+        self._mult_references += mult_references