+++ src/aac_metrics/experimental/mult_cands.py
@@ -10,24 +10,30 @@ from torch import Tensor
 
 def mult_cands_wrapper(
     metric: Callable,
+    metric_out_name: str,
     mult_candidates: list[list[str]],
     mult_references: list[list[str]],
-    reduction: Union[str, Callable[[Tensor], Tensor]] = "max",
-    *args,
+    return_all_scores: bool = True,
+    selection: str = "max",
+    reduction: Callable[[Tensor], Tensor] = torch.mean,
     **kwargs,
-) -> tuple[dict[str, Tensor], dict[str, Tensor]]:
+) -> Union[Tensor, tuple[dict[str, Tensor], dict[str, Tensor]]]:
     """Multiple candidates metric wrapper.
 
     :param metric: Any Callable metric code.
+    :param metric_out_name: The name of the metric output.
     :param mult_candidates: The candidates input. Currently only supports having the same number of multiple candidates.
     :param mult_references: The references input.
-    :param reduction: The reduction function to apply. Take Tensor of shape (n_cand_per_audio, n_audio) and reduce to (n_audio,).
-        defaults to "max".
-    :param metric_prefix: The prefix added to the output metric names. defaults to "mcands.".
-    :param metric_suffix: The suffix added to the output metric names. defaults to ".max".
-    :param *args: The positional arguments given to the metric call.
+    :param selection: The selection to apply. Can be "max", "min" or "mean". defaults to "max".
+    :param reduction: The reduction function to apply to local scores. defaults to torch.mean.
     :param **kwargs: The keywords arguments given to the metric call.
     """
+    SELECTIONS = ("max", "min", "mean")
+    if selection not in SELECTIONS:
+        raise ValueError(
+            f"Invalid argument {selection=}. (expected one of {SELECTIONS})"
+        )
+
     if len(mult_candidates) <= 0:
         raise ValueError(
             f"Cannot compute max metric without at least 1 hypothesis. (found {len(mult_candidates)=})"
@@ -43,30 +49,45 @@ def mult_cands_wrapper(
             "Cannot compute multiple candidates metric with a various number of candidates."
         )
 
-    if isinstance(reduction, str):
-        if reduction == "max":
-            reduction = _max_reduce
-        else:
-            raise ValueError(f"Invalid argument {reduction=}.")
-
     all_local_scores_lst: list[dict[str, Tensor]] = []
 
     for i in range(n_cands_per_audio):
         candidates_i = [cands[i] for cands in mult_candidates]
         _global_scores_i, local_scores_i = metric(
-            candidates_i, mult_references, *args, return_all_scores=True, **kwargs
+            candidates_i,
+            mult_references,
+            return_all_scores=True,
+            **kwargs,
         )
         all_local_scores_lst.append(local_scores_i)
 
+    # list[dict[str, Tensor]] to dict[str, stacked Tensor]
     keys = list(all_local_scores_lst[0].keys())
     all_local_scores = {
         k: torch.stack([local_scores_i[k] for local_scores_i in all_local_scores_lst])
         for k in keys
     }
-    local_scores = {k: reduction(scores) for k, scores in all_local_scores.items()}
-    global_scores = {k: scores.mean() for k, scores in local_scores.items()}
-    return local_scores, global_scores
 
+    if selection == "max":
+        indexes = all_local_scores[metric_out_name].argmax(dim=0)
+        local_scores = {}
 
-def _max_reduce(x: Tensor) -> Tensor:
-    return x.max(dim=0).values
+    elif selection == "min":
+        indexes = all_local_scores[metric_out_name].argmin(dim=0)
+        local_scores = {k: scores[indexes] for k, scores in all_local_scores.items()}
+
+    elif selection == "mean":
+        selected_scores = all_local_scores[metric_out_name].mean(dim=0)
+        local_scores = {metric_out_name: selected_scores}
+
+    else:
+        raise ValueError(
+            f"Invalid argument {selection=}. (expected one of {SELECTIONS})"
+        )
+
+    global_scores = {k: reduction(scores) for k, scores in local_scores.items()}
+
+    if return_all_scores:
+        return local_scores, global_scores
+    else:
+        return global_scores[metric_out_name]