+++ CHANGELOG.md
@@ -4,4 +4,5 @@ All notable changes to this project will be documented in this file.
 
 ## [0.1.0] UNRELEASED
 ### Added
-- Bleu, METEOR, ROUGE-l, SPICE, CIDEr and SPIDEr metrics.
+- BLEU, METEOR, ROUGE-l, SPICE, CIDEr and SPIDEr metrics functions and modules.
+- SPIDEr-max experimental implementation.